pyproject.toml
[project]
name = "obsidian-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index>=0.12.34",
    "llama-index-core>=0.12.34.post1",
    "llama-index-embeddings-huggingface>=0.5.3",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-readers-obsidian==0.5.1",
    "llama-index-vector-stores-qdrant>=0.6.0",
    "qdrant-client>=1.14.2",
    "python-dotenv>=1.0.0",
]
